@@ -8,7 +8,7 @@ name = "versiontracker"
 dynamic = ["version"]
 description = "CLI versiontracker and update tool for macOS"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = {text = "MIT"}
 authors = [
     {name = "docdyhr", email = "thomas@dyhr.com"}
@@ -22,7 +22,6 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: MacOS :: MacOS X",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -32,10 +31,13 @@ classifiers = [
 ]
 dependencies = [
     "fuzzywuzzy>=0.18.0",
-    "python-Levenshtein>=0.12.2",
     "rapidfuzz>=3.0.0",
     "tqdm>=4.66.0",
     "PyYAML>=6.0",
+    "termcolor>=2.3.0",
+    "tabulate>=0.9.0",
+    "psutil>=5.9.5",
+    "aiohttp>=3.8.0",
 ]
 
 [project.urls]
@@ -50,8 +52,8 @@ versiontracker = "versiontracker.__main__:main"
 
 [tool.ruff]
 # Define the target Python version.
-# We support Python 3.8 and later.
-target-version = "py38"
+# We support Python 3.9 and later.
+target-version = "py39"
 
 # Exclude the venv directory.
 exclude = [
