@@ -8,8 +8,8 @@ import asyncio
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
-from aiohttp import ClientSession
 
+from tests.mock_aiohttp_session import create_mock_session_factory, mock_aiohttp_session
 from versiontracker.async_homebrew import (
     HomebrewBatchProcessor,
     HomebrewVersionChecker,
@@ -22,32 +22,17 @@ from versiontracker.async_homebrew import (
 from versiontracker.exceptions import NetworkError
 
 
-@pytest.fixture
-def mock_aiohttp_session():
-    """Create a mock aiohttp session."""
-    session = MagicMock(spec=ClientSession)
-    response = MagicMock()
-    session.get.return_value.__aenter__.return_value = response
-    response.raise_for_status = MagicMock()
-    response.json.return_value = asyncio.Future()
-    response.json.return_value.set_result({"name": "firefox", "version": "100.0"})
-    return session, response
-
-
 @pytest.mark.asyncio
-async def test_fetch_cask_info_success(mock_aiohttp_session):
+async def test_fetch_cask_info_success():
     """Test successful cask info fetching."""
-    session, response = mock_aiohttp_session
+    with patch(
+        "versiontracker.async_homebrew.fetch_json",
+        return_value={"name": "firefox", "version": "100.0"},
+    ):
+        result = await fetch_cask_info("firefox", use_cache=True)
 
-    with patch("aiohttp.ClientSession", return_value=session):
-        with patch(
-            "versiontracker.async_homebrew.fetch_json",
-            return_value={"name": "firefox", "version": "100.0"},
-        ):
-            result = await fetch_cask_info("firefox", use_cache=True)
-
-            # Verify the result
-            assert result == {"name": "firefox", "version": "100.0"}
+        # Verify the result
+        assert result == {"name": "firefox", "version": "100.0"}
 
 
 @pytest.mark.asyncio
@@ -67,25 +52,26 @@ async def test_search_casks_success():
     # Mock search results
     mock_results = [
         {"token": "firefox-cask", "name": "Firefox"},
-        {"token": "not-firefox", "name": "Not Firefox"},
+        {"token": "another-cask", "name": "Another App"},
     ]
 
-    with patch("versiontracker.async_homebrew.read_cache", return_value=None):
-        with patch("aiohttp.ClientSession") as mock_session:
-            # Setup the mock response
-            mock_response = AsyncMock()
-            mock_response.raise_for_status = AsyncMock()
-            mock_response.json.return_value = mock_results
+    async with mock_aiohttp_session() as mock_session:
+        # Setup mock response
+        mock_session.add_response(
+            "https://formulae.brew.sh/api/search.json?q=firefox", json_data=mock_results
+        )
 
-            # Setup the mock session
-            mock_session.return_value.__aenter__.return_value.get.return_value.__aenter__.return_value = mock_response
+        session_factory = create_mock_session_factory(mock_session)
 
+        with patch("versiontracker.async_homebrew.read_cache", return_value=None):
             with patch("versiontracker.async_homebrew.write_cache", return_value=True):
-                results = await search_casks("firefox", use_cache=False)
+                with patch("aiohttp.ClientSession", session_factory):
+                    results = await search_casks("firefox", use_cache=False)
 
-                # Verify the results
-                assert len(results) == 2
-                assert results[0]["token"] == "firefox-cask"
+                    # Verify the results
+                    assert len(results) == 2
+                    assert results[0]["token"] == "firefox-cask"
+                    assert results[1]["token"] == "another-cask"
 
 
 @pytest.mark.asyncio
@@ -184,28 +170,31 @@ class TestHomebrewBatchProcessor:
         assert processor._is_significant_match("A", "verylongname") is False
 
 
-@pytest.mark.asyncio
-async def test_async_get_cask_version():
+def test_async_get_cask_version():
     """Test getting a cask version."""
-    # Mock the fetch_cask_info function
+    # Mock the fetch_cask_info function to return a coroutine
     with patch(
         "versiontracker.async_homebrew.fetch_cask_info",
+        new_callable=AsyncMock,
         return_value={"version": "100.0"},
     ):
-        version = await async_get_cask_version("firefox")
+        version = async_get_cask_version("firefox")
 
         # Verify the version
         assert version == "100.0"
 
 
-@pytest.mark.asyncio
-async def test_async_get_cask_version_not_found():
+def test_async_get_cask_version_not_found():
     """Test getting a version for a non-existent cask."""
     # Mock fetch_cask_info to raise a 404 error
     mock_error = NetworkError("HTTP error 404: Not Found")
 
-    with patch("versiontracker.async_homebrew.fetch_cask_info", side_effect=mock_error):
-        version = await async_get_cask_version("nonexistent-cask")
+    with patch(
+        "versiontracker.async_homebrew.fetch_cask_info",
+        new_callable=AsyncMock,
+        side_effect=mock_error,
+    ):
+        version = async_get_cask_version("nonexistent-cask")
 
         # Verify the version is None
         assert version is None
@@ -221,7 +210,9 @@ class TestHomebrewVersionChecker:
 
         # Mock the async_get_cask_version function
         with patch(
-            "versiontracker.async_homebrew.async_get_cask_version", return_value="101.0"
+            "versiontracker.async_homebrew.async_get_cask_version",
+            new_callable=AsyncMock,
+            return_value="101.0",
         ):
             result = await checker.process_item(("Firefox", "100.0", "firefox"))
 
@@ -244,7 +235,8 @@ class TestHomebrewVersionChecker:
             assert result == ("Firefox", "100.0", "firefox", None)
 
 
-def test_async_check_brew_install_candidates():
+@pytest.mark.asyncio
+async def test_async_check_brew_install_candidates():
     """Test checking brew install candidates."""
     # Mock the process_all method
     with patch(
@@ -253,26 +245,44 @@ def test_async_check_brew_install_candidates():
         with patch.object(
             HomebrewBatchProcessor,
             "process_all",
+            new_callable=AsyncMock,
             return_value=[("Firefox", "100.0", True)],
         ):
-            results = async_check_brew_install_candidates([("Firefox", "100.0")])
+            # Access the underlying async function without the sync wrapper
+            if hasattr(async_check_brew_install_candidates, "__wrapped__"):
+                results = await async_check_brew_install_candidates.__wrapped__(
+                    [("Firefox", "100.0")]
+                )
+            else:
+                results = await async_check_brew_install_candidates(
+                    [("Firefox", "100.0")]
+                )
 
             # Verify the results
             assert results == [("Firefox", "100.0", True)]
 
 
-def test_async_check_brew_install_candidates_no_homebrew():
+@pytest.mark.asyncio
+async def test_async_check_brew_install_candidates_no_homebrew():
     """Test checking brew install candidates when Homebrew is not available."""
     with patch(
         "versiontracker.async_homebrew.is_homebrew_available", return_value=False
     ):
-        results = async_check_brew_install_candidates([("Firefox", "100.0")])
+        # Access the underlying async function without the sync wrapper
+        if hasattr(async_check_brew_install_candidates, "__wrapped__"):
+            results = await async_check_brew_install_candidates.__wrapped__(
+                [("Firefox", "100.0")]
+            )
+        else:
+            results = await async_check_brew_install_candidates([("Firefox", "100.0")])
 
         # Verify the results
         assert results == [("Firefox", "100.0", False)]
+        # Verify the results
 
 
-def test_async_check_brew_update_candidates():
+@pytest.mark.asyncio
+async def test_async_check_brew_update_candidates():
     """Test checking brew update candidates."""
     # Mock the process_all method
     with patch(
@@ -281,22 +291,38 @@ def test_async_check_brew_update_candidates():
         with patch.object(
             HomebrewVersionChecker,
             "process_all",
+            new_callable=AsyncMock,
             return_value=[("Firefox", "100.0", "firefox", "101.0")],
         ):
-            results = async_check_brew_update_candidates(
-                [("Firefox", "100.0", "firefox")]
-            )
+            # Access the underlying async function without the sync wrapper
+            if hasattr(async_check_brew_update_candidates, "__wrapped__"):
+                results = await async_check_brew_update_candidates.__wrapped__(
+                    [("Firefox", "100.0", "firefox")]
+                )
+            else:
+                results = await async_check_brew_update_candidates(
+                    [("Firefox", "100.0", "firefox")]
+                )
 
             # Verify the results
             assert results == [("Firefox", "100.0", "firefox", "101.0")]
 
 
-def test_async_check_brew_update_candidates_no_homebrew():
+@pytest.mark.asyncio
+async def test_async_check_brew_update_candidates_no_homebrew():
     """Test checking brew update candidates when Homebrew is not available."""
     with patch(
         "versiontracker.async_homebrew.is_homebrew_available", return_value=False
     ):
-        results = async_check_brew_update_candidates([("Firefox", "100.0", "firefox")])
+        # Access the underlying async function without the sync wrapper
+        if hasattr(async_check_brew_update_candidates, "__wrapped__"):
+            results = await async_check_brew_update_candidates.__wrapped__(
+                [("Firefox", "100.0", "firefox")]
+            )
+        else:
+            results = await async_check_brew_update_candidates(
+                [("Firefox", "100.0", "firefox")]
+            )
 
         # Verify the results
         assert results == [("Firefox", "100.0", "firefox", None)]
