@@ -136,7 +136,7 @@ def test_get_version_difference(version1, version2, expected_difference):
     "current_version,latest_version,expected_status",
     [
         # Exact same versions
-        ("1.0.0", "1.0.0", VersionStatus.UPTODATE),
+        ("1.0.0", "1.0.0", VersionStatus.UP_TO_DATE),
         # Outdated versions
         ("1.0.0", "1.0.1", VersionStatus.OUTDATED),
         ("1.0.0", "1.1.0", VersionStatus.OUTDATED),
@@ -186,18 +186,24 @@ def test_version_difference_components(
 ):
     """Test individual component differences in versions."""
     version_info = get_version_info(version1, version2)
-    if version_info.status != VersionStatus.UNKNOWN:
+    if version_info.status == VersionStatus.UP_TO_DATE:
+        # For equal versions, expect zero differences
+        assert expected_major == 0 and expected_minor == 0 and expected_patch == 0
+    elif version_info.status != VersionStatus.UNKNOWN:
         diff = (
             version_info.outdated_by
             if version_info.status == VersionStatus.OUTDATED
             else version_info.newer_by
         )
-        assert diff[0] == expected_major, (
-            f"Expected major diff {expected_major} but got {diff[0]}"
+        assert diff is not None, (
+            f"Expected difference but got None for status {version_info.status}"
         )
-        assert diff[1] == expected_minor, (
-            f"Expected minor diff {expected_minor} but got {diff[1]}"
+        assert diff[0] == abs(expected_major), (
+            f"Expected major diff {abs(expected_major)} but got {diff[0]}"
         )
-        assert diff[2] == expected_patch, (
-            f"Expected patch diff {expected_patch} but got {diff[2]}"
+        assert diff[1] == abs(expected_minor), (
+            f"Expected minor diff {abs(expected_minor)} but got {diff[1]}"
+        )
+        assert diff[2] == abs(expected_patch), (
+            f"Expected patch diff {abs(expected_patch)} but got {diff[2]}"
         )
