@@ -716,14 +716,11 @@ class AdvancedCache:
         # Try to get from cache
         cached_value = self.get(key, level, ttl)
         if cached_value is not None:
-            return cached_value
+            return cached_value  # type: ignore
 
-        # Not in cache, call getter function
+        # Not in cache, call getter function and store it
         value = getter_func()
-
-        # Store in cache
         self.put(key, value, level, priority, source)
-
         return value
 
     def batch_get(
